+++ typescript/keyfunder/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @hyperlane-xyz/keyfunder
 
+## 0.1.19
+
+### Patch Changes
+
+- Updated dependencies [d5168fc]
+  - @hyperlane-xyz/utils@31.0.1
+  - @hyperlane-xyz/core@11.3.1
+  - @hyperlane-xyz/metrics@0.2.12
+  - @hyperlane-xyz/sdk@31.0.1
+
 ## 0.1.18
 
 ### Patch Changes