+++ typescript/tron-sdk/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @hyperlane-xyz/tron-sdk
 
+## 22.1.11
+
+### Patch Changes
+
+- Updated dependencies [d5168fc]
+  - @hyperlane-xyz/utils@31.0.1
+  - @hyperlane-xyz/core@11.3.1
+  - @hyperlane-xyz/provider-sdk@4.3.1
+
 ## 22.1.10
 
 ### Patch Changes